requirements.txt
pillow==7.2.0
youtube_dl==2020.9.20